+++ CHANGELOG.md
@@ -10,8 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - Task grouped counts added to "inspect queue" in cli tool
+- Can purge selected tasks from queue with the cli tool, by task name, task id or app name
 - Can purge task logs with cli tool
-- Can purge selected tasks from queue with the cli tool, e.g. by task name and app name
 
 ### Fixed